refactor: rename typo "contants" to "constants"
